setloader.pyi
Took 3 minutes
